Versi 1 (raw, no js n css)/requirements.txt
Flask==2.0.2
mysql-connector-python==8.2.0
